test(315): silent-failure-hunter C1/C2 対応 — selector 拡張 + 型契約 test 分離
PR #329 の 6 エージェント並列レビューで検出された Critical 2 件への対応:

C1 (silent-failure-hunter): ESLint selector が dot-access しか検知せず、以下の bypass path が
silent PASS 経路になっていた:
- process.env['NODE_ENV'] = ... (computed property, bracket literal)
- Object.assign(process.env, { NODE_ENV: ... }) (bulk mutation)

対応: selector を 3 パターンに拡張し、それぞれ専用エラーメッセージを設定。
dummy violation 3 種 (dot / bracket / Object.assign) で全て EXIT 1 + 日本語メッセージ発火を確認。
dynamic key (process.env[key] = ...) は selector で原理的に拾えないが、現実的な bypass path は
本 rule で覆う。

C2 (silent-failure-hunter): tsconfig.test.json の include が test/types/**/*.ts のみで、
test/helpers/withNodeEnv.test.ts の @ts-expect-error 型契約 test が
`npm run type-check:test` では silent PASS する。npm test の ts-node/register 経由でのみ
検知されるため、CI の type-check:test だけ走る経路で失効。

対応: 型契約 describe を既存規約に沿い test/types/withNodeEnv.types.test.ts に分離。
tsconfig.test.json の include に含まれ、npm run type-check:test で strict 検査される。
既存 test/types/ の pageOcrResult.types.test.ts / textCapGenericsContract.test.ts 等と
同一 convention (describe/it + @ts-expect-error + test-strategy.md §2.2 参照)。

comment-analyzer Important 対応: helper override glob (test/helpers/**/*.ts) が広い旨を
コメントで明示、将来 env-mutation helper 追加時のレビュー注意点を記載。

テスト: 610 passing 維持 (型契約 3 ケースは helpers/ から types/ に移動、総数同じ)
tsc/type-check:test: EXIT 0
eslint: dummy violation 3 種全て EXIT 1、helpers/ 除外 OK、既存 warning 1 件のみ (本 PR 無関係)
